seperate out Metadata component from Head
