+++ frontend/components/common/index.js
@@ -4,6 +4,7 @@ export * from './EmptyState'
 export * from './Grid'
 export * from './Icon'
 export * from './Line'
+export * from './Metadata'
 export * from './Modal'
 export * from './Loading'
 export * from './ProfilePic'
